refactor: Model 폴더로 게임 내부 동작 클래스 이동
